Removing private stimuli and extra_data functions
