+++ src/extra_data_library/extra_data.c
@@ -7,47 +7,9 @@
 #include "../config/extra_data_config.h"
 #include "../config_helpers/config_helpers.h"
 #include "../libraries_common/common_data_structures.h"
-#include "../utils/file_utils.h"
 #include "../domains_library/mesh_info_data.h"
+#include "helper_functions.h"
 
-
-real* set_common_schemia_data(struct config *config, uint32_t num_cells, int num_par, size_t *extra_data_size) {
-
-    *extra_data_size = sizeof(real)*(num_cells + num_par);
-
-    real *extra_data = (real*)malloc(*extra_data_size);
-
-    real atpi = 6.8;
-    GET_PARAMETER_NUMERIC_VALUE_OR_USE_DEFAULT(real, atpi, config, "atpi");
-
-    real Ko = 5.4;
-    GET_PARAMETER_NUMERIC_VALUE_OR_USE_DEFAULT(real, Ko, config, "Ko");
-
-    real Ki = 138.3;
-    GET_PARAMETER_NUMERIC_VALUE_OR_USE_DEFAULT(real, Ki, config, "Ki");
-
-    real GNa_multiplicator = 1.0f;
-    GET_PARAMETER_NUMERIC_VALUE_OR_USE_DEFAULT(real, GNa_multiplicator, config, "GNa_multiplicator");
-
-    real GCaL_multiplicator = 1.0f;
-    GET_PARAMETER_NUMERIC_VALUE_OR_USE_DEFAULT(real, GCaL_multiplicator, config, "GCaL_multiplicator");
-
-    real INaCa_multiplicator = 1.0f;
-    GET_PARAMETER_NUMERIC_VALUE_OR_USE_DEFAULT(real, INaCa_multiplicator, config, "INaCa_multiplicator");
-
-    real Vm_modifier = 0.0f;
-    GET_PARAMETER_NUMERIC_VALUE_OR_USE_DEFAULT(real, Vm_modifier, config, "Vm_modifier");
-
-    extra_data[0] = atpi;
-    extra_data[1] = Ko;
-    extra_data[2] = Ki;
-    extra_data[3] = Vm_modifier;
-    extra_data[4] = GNa_multiplicator;
-    extra_data[5] = GCaL_multiplicator;
-    extra_data[6] = INaCa_multiplicator;
-
-    return extra_data;
-}
 
 SET_EXTRA_DATA(set_extra_data_for_fibrosis_sphere) {
 
@@ -127,209 +89,6 @@ SET_EXTRA_DATA(set_extra_data_for_no_fibrosis) {
     return (void*)fibs;
 }
 
-SET_EXTRA_DATA(set_extra_data_for_human_full_mesh) {
-
-    uint32_t num_active_cells = the_grid->num_active_cells;
-
-    int num_par = 7;
-    real *fibs = NULL;
-    fibs = set_common_schemia_data(config, num_active_cells, num_par, extra_data_size);
-
-    struct cell_node ** ac = the_grid->active_cells;
-
-    real_cpu small_scar_center_x = 0.0;
-    GET_PARAMETER_NUMERIC_VALUE_OR_REPORT_ERROR(real, small_scar_center_x, config, "small_scar_center_x");
-
-    real_cpu small_scar_center_y = 0.0;
-    GET_PARAMETER_NUMERIC_VALUE_OR_REPORT_ERROR(real, small_scar_center_y, config, "small_scar_center_y");
-
-    real_cpu small_scar_center_z = 0.0;
-    GET_PARAMETER_NUMERIC_VALUE_OR_REPORT_ERROR(real, small_scar_center_z, config, "small_scar_center_z");
-
-    real_cpu big_scar_center_x = 0.0;
-    GET_PARAMETER_NUMERIC_VALUE_OR_REPORT_ERROR(real, big_scar_center_x, config, "big_scar_center_x");
-
-    real_cpu big_scar_center_y = 0.0;
-    GET_PARAMETER_NUMERIC_VALUE_OR_REPORT_ERROR(real, big_scar_center_y, config, "big_scar_center_y");
-
-    real_cpu big_scar_center_z = 0.0;
-    GET_PARAMETER_NUMERIC_VALUE_OR_REPORT_ERROR(real, big_scar_center_z, config, "big_scar_center_z");
-
-    real_cpu bz_size_big = 0;
-    real_cpu bz_size_small = 0;
-    real_cpu dist_big = 0;
-    real_cpu dist_small = 0;
-
-	uint32_t i;
-	bool fibrotic, border_zone;
-	int scar_type;
-
-	OMP(parallel for private(dist_big, dist_small))
-    for (i = 0; i < num_active_cells; i++) {
-
-        border_zone = DHZB_MESH_TISSUE_TYPE(ac[i]) == BZ;
-        scar_type = DHZB_MESH_LOCATION(ac[i]);
-
-        if (ac[i]->active && border_zone) {
-            real_cpu center_x = ac[i]->center.x;
-            real_cpu center_y = ac[i]->center.y;
-            real_cpu center_z = ac[i]->center.z;
-            if(scar_type == BIG_SCAR) {
-                dist_big = sqrt((center_x - big_scar_center_x) * (center_x - big_scar_center_x) +
-                                (center_y - big_scar_center_y) * (center_y - big_scar_center_y) +
-                                (center_z - big_scar_center_z) * (center_z - big_scar_center_z));
-                OMP(critical(big))
-                if (dist_big > bz_size_big) {
-                    bz_size_big = dist_big;
-                }
-            }
-            else if(scar_type == SMALL_SCAR) {
-                dist_small = sqrt((center_x - small_scar_center_x) * (center_x - small_scar_center_x) +
-                                  (center_y - small_scar_center_y) * (center_y - small_scar_center_y) +
-                                  (center_z - small_scar_center_z) * (center_z - small_scar_center_z));
-                OMP(critical(small))
-                if (dist_small > bz_size_small) {
-                    bz_size_small = dist_small;
-                }
-            }
-        }
-    }
-
-    OMP(parallel for private(dist_big, dist_small))
-    for (i = 0; i < num_active_cells; i++) {
-
-        if (ac[i]->active) {
-            fibrotic = DHZB_MESH_TISSUE_TYPE(ac[i]) == SCAR;
-            border_zone = DHZB_MESH_TISSUE_TYPE(ac[i]) == BZ;
-            scar_type = DHZB_MESH_LOCATION(ac[i]);
-
-            if(fibrotic) {
-                fibs[i+num_par] = 0.0f;
-            }
-            else if (border_zone) {
-                real_cpu center_x = ac[i]->center.x;
-                real_cpu center_y = ac[i]->center.y;
-                real_cpu center_z = ac[i]->center.z;
-                if(scar_type == BIG_SCAR) {
-                    dist_big = sqrt((center_x - big_scar_center_x) * (center_x - big_scar_center_x) +
-                                    (center_y - big_scar_center_y) * (center_y - big_scar_center_y) +
-                                    (center_z - big_scar_center_z) * (center_z - big_scar_center_z));
-                    fibs[i+num_par] = (real)(dist_big / bz_size_big);
-
-                }
-                else if(scar_type == SMALL_SCAR) {
-                    dist_small = sqrt((center_x - small_scar_center_x) * (center_x - small_scar_center_x) +
-                                      (center_y - small_scar_center_y) * (center_y - small_scar_center_y) +
-                                      (center_z - small_scar_center_z) * (center_z - small_scar_center_z));
-                    fibs[i+num_par] = (real)(dist_small / bz_size_small);
-                }
-            }
-            else {
-                fibs[i+num_par] = 1.0f;
-            }
-        }
-    }
-
-    return (void*)fibs;
-}
-
-SET_EXTRA_DATA(set_extra_data_for_scar_wedge) {
-
-    uint32_t num_active_cells = the_grid->num_active_cells;
-    real *fibs = NULL;
-
-    int num_par = 7;
-    fibs = set_common_schemia_data(config, num_active_cells, num_par, extra_data_size);
-
-    struct cell_node ** ac = the_grid->active_cells;
-
-    char *scar_size;
-    GET_PARAMETER_STRING_VALUE_OR_REPORT_ERROR (scar_size, config, "scar_size");
-
-    uint8_t size_code;
-
-    if(strcmp(scar_size, "big") == 0) {
-        size_code = 0;
-    }
-    else if(strcmp(scar_size, "small") == 0) {
-        size_code = 1;
-    }
-    else {
-        printf("Function: set_extra_data_for_scar_edge, invalid scar size %s. Valid sizes are big or small. Exiting!\n", scar_size);
-        exit(EXIT_FAILURE);
-    }
-
-    real_cpu scar_center_x;
-    real_cpu scar_center_y;
-    real_cpu scar_center_z;
-
-    ////Fibrosis configuration
-    //BIG SCAR
-    if(size_code == 0) {
-        scar_center_x = 95300;
-        scar_center_y = 81600;
-        scar_center_z = 36800;
-    }
-    else {
-        scar_center_x = 52469;
-        scar_center_y = 83225;
-        scar_center_z = 24791;
-    }
-    real_cpu bz_size = 0.0;
-    real_cpu dist;
-
-	uint32_t i;
-	bool border_zone, fibrotic;
-
-	OMP(parallel for private(dist))
-    for (i = 0; i < num_active_cells; i++) {
-        if(ac[i]->active) {
-            border_zone = DHZB_MESH_TISSUE_TYPE(ac[i]) == BZ;
-            if(border_zone) {
-                real_cpu center_x = ac[i]->center.x;
-                real_cpu center_y = ac[i]->center.y;
-                real_cpu center_z = ac[i]->center.z;
-                dist =  sqrt((center_x - scar_center_x)*(center_x - scar_center_x) + (center_y - scar_center_y)*(center_y - scar_center_y)  + (center_z - scar_center_z)*(center_z - scar_center_z)  );
-                OMP(critical)
-                if(dist > bz_size) {
-                    bz_size = dist;
-                }
-            }
-
-        }
-    }
-
-    OMP(parallel for private(dist))
-    for (i = 0; i < num_active_cells; i++) {
-
-        if(ac[i]->active) {
-
-            fibrotic = DHZB_MESH_TISSUE_TYPE(ac[i]) == SCAR;
-            border_zone = DHZB_MESH_TISSUE_TYPE(ac[i]) == BZ;
-
-            if(fibrotic) {
-                fibs[i+num_par] = 0.0;
-            }
-            else if(border_zone) {
-                real_cpu center_x = ac[i]->center.x;
-                real_cpu center_y = ac[i]->center.y;
-                real_cpu center_z = ac[i]->center.z;
-                dist =  sqrt((center_x - scar_center_x)*(center_x - scar_center_x) + (center_y - scar_center_y)*(center_y - scar_center_y)  + (center_z - scar_center_z)*(center_z - scar_center_z)  );
-                dist = dist/bz_size;
-
-                fibs[i + num_par] = (real)dist;
-
-            }
-            else {
-                fibs[i + num_par] = 1.0f;
-            }
-
-        }
-    }
-
-    return (void*)fibs;
-}
-
 SET_EXTRA_DATA(set_extra_data_for_benchmark) {
 
     *extra_data_size = sizeof(real)*19;
@@ -358,27 +117,6 @@ SET_EXTRA_DATA(set_extra_data_for_benchmark) {
     initial_conditions[18] = 138.52;   // K_i;     millimolar;    138.52
 
     return (void*)initial_conditions;
-}
-
-SET_EXTRA_DATA(set_extra_data_for_scv_mesh) {
-
-	uint32_t num_active_cells = the_grid->num_active_cells;
-
-	*extra_data_size = sizeof(uint32_t)*(num_active_cells);
-
-	uint32_t *mapping = (uint32_t*)malloc(*extra_data_size);
-
-	struct cell_node ** ac = the_grid->active_cells;
-
-	int i;
-
-	OMP(parallel for)
-	for (i = 0; i < num_active_cells; i++) {
-		mapping[i] = TISSUE_TYPE(ac[i]); //endo 0; mid  1; epi  2
-	}
-
-	return (void*)mapping;
-
 }
 
 SET_EXTRA_DATA (set_mixed_model_if_x_less_than)
@@ -439,4 +177,3 @@ SET_EXTRA_DATA (set_mixed_model_purkinje_and_tissue)
 
     return (void*)mapping;
 }
-
