+++ src/extra_data_library/helper_functions.h
@@ -1,0 +1,15 @@
+//
+// Created by sachetto on 19/03/2021.
+//
+
+#ifndef MONOALG3D_C_EXTRA_DATA_HELPER_FUNCTIONS_H
+#define MONOALG3D_C_EXTRA_DATA_HELPER_FUNCTIONS_H
+
+#include <unistd.h>
+#include "../common_types/common_types.h"
+#include "../config/config_common.h"
+
+
+real* set_common_schemia_data(struct config *config, uint32_t num_cells, int num_par, size_t *extra_data_size);
+
+#endif // MONOALG3D_C_EXTRA_DATA_HELPER_FUNCTIONS_H
